Changes to prevent Heroku from inferring a python buildpack.
Robot E2E tests use python, but don't need to run on Heroku.  So we should just install the dependencies in circle-ci instead.

The presence of the robot `requirements.txt` file was causing Heroku to think this is a python app; I've moved the equivalent commands into `circle.yml` instead.